remove dangling logseq refs from calendar
